% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertools.R
\docType{package}
\name{advertools}
\alias{advertools}
\alias{advertools-package}
\title{advertools: Productivity and analysis tools for online marketing.}
\description{
A package for helping online marketing people become more productive
and gain better insights from their campaigns and data. Several functions
help automate and make efficient various tasks such as generating keywords,
analyzing weighted word frequency, tracking competition and more.
}

