% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_frequency.R
\docType{data}
\name{boxoffice}
\alias{boxoffice}
\title{The boxoffice revenues of the top 100 movies of all time}
\format{A data frame with 101 rows and 2 variables:
\describe{
   \item{Movie_Title}{Name of the movie}
   \item{Lifetime.Gross}{Lifetime revnue in US$}
}}
\source{
\url{http://www.boxofficemojo.com/alltime/domestic.htm}
}
\usage{
boxoffice
}
\description{
A dataset of the top 100 grossing movies of all time, as of April
2016.
}
\keyword{datasets}

