% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw_combinations.R
\name{kw_combinations}
\alias{kw_combinations}
\title{Keyword combinations}
\usage{
kw_combinations(df, cols, sep = " ")
}
\arguments{
\item{df}{data frame containing different word variations in each column}

\item{cols}{character vector of names of the columns that you want to
combine. The order is important here}

\item{sep}{Optional separator of the final keywords generated}
}
\value{
data fram with a column for each keywrod template
}
\description{
This function helps you generate combinations and sequences of words, to
obtain meaningful keywords for your campaigns. It assumes you already have a
data frame with different words for each element of search
}

