% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{kw_exact}
\alias{kw_exact}
\title{Exact match keywords}
\usage{
kw_exact(x)
}
\arguments{
\item{x}{a keyword or a vector of keywords}
}
\value{
keywords in exact match
}
\description{
Wraps square brackets around a character vector to turn them into exact-match
keywords. It takes care of removing any other match characters like +, -, or ".
}
\examples{
mtcars$exact <- kw_exact(rownames(mtcars))
keywords <- c("movie tickets", "movies", "best movies")
kw_exact(keywords)
}

