% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{kw_negative}
\alias{kw_negative}
\title{Negative match keywords}
\usage{
kw_negative(x)
}
\arguments{
\item{x}{a keyword or a vector of keywords}
}
\value{
negative match keywords
}
\description{
Adds a minus sign to each word in a character vector to turn them into negative match
keywords. It takes care of removing any other match characters like +, [, or ".
}
\examples{
mtcars$negative <- kw_negative(rownames(mtcars))
keywords <- c("movie tickets", "movies", "best movies")
kw_negative(keywords)
}

