% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{kw_negative_phrase}
\alias{kw_negative_phrase}
\title{Negative phrase match keywords}
\usage{
kw_negative_phrase(x)
}
\arguments{
\item{x}{a keyword or a vector of keywords}
}
\value{
negative phrase match keywords
}
\description{
Wraps a minus sign as well as quotation marks around a character vector to turn them into
the "embedded" negative phrase match keywords.
It takes care of removing any other match characters like +, -, or ".
}
\examples{
mtcars$negative_phrase <- kw_negative_phrase(rownames(mtcars))
keywords <- c("movie tickets", "movies", "best movies")
kw_negative_phrase(keywords)
}

