% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{kw_phrase}
\alias{kw_phrase}
\title{Phrase match keywords}
\usage{
kw_phrase(x)
}
\arguments{
\item{x}{a keyword or a vector of keywords}
}
\value{
keywords in phrase match
}
\description{
Wraps quotation marks around a character vector to turn them into phrase-match
keywords. It takes care of removing any other match characters like +, -, or [
if they exist.
}
\examples{
mtcars$phrase <- kw_phrase(rownames(mtcars))
keywords <- c("movie tickets", "movies", "best movies")
kw_phrase(keywords)
}

