% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_search_ads}
\alias{sample_search_ads}
\title{A sample result of search ads}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 27 rows and 7 columns.}
\usage{
sample_search_ads
}
\description{
A few ads retrieved through the search_ads_google() function, to illustrate
how it looks like and for demonstration purposes.
}
\keyword{datasets}

