% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ads_google.R
\name{search_ads_google}
\alias{search_ads_google}
\title{Search Ads - Google}
\usage{
search_ads_google(keywords, domains = "com")
}
\arguments{
\item{keywords}{a keyword or a character vector}

\item{domains}{the Google domain you wish to check}
}
\value{
data frame with ad text with some metadata
}
\description{
Analyze the competition or use it for explaring who is advertising for what.
Enter a keyword or a vector of keywords, specify the Google domain
(com, co.jp, etc), and get a data frame showing keywords, ads, and the advertisers
who are bidding on this keyword at this particular moment.
This is valuable in seeing how you fare compared to competitors. It also helps you
figure out strategies of other advertisers and whether they are consistent, changing
strategies, specific, aggressive, etc.
}
\details{
For long ongoing campaigns, ideally you can run a report for your top keywords a few
times a day, and bind the data in one large data frame. This can augment your campaign
data in interesting ways and can give you additional insights.
}
\examples{
search_ads_google(c("flights to new york", "cheap flights", "nyc airline tickets"))
}

