% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_ar}
\alias{stopwords_ar}
\title{Arabic stopwords}
\format{An object of class \code{character} of length 107.}
\source{
\href{http://www.ranks.nl/stopwords/arabic}{Ranks.nl}
}
\usage{
stopwords_ar
}
\description{
107 words frequenty used Arabic words for removal in text mining,
and based on the list adapted from ranks.nl
}
\keyword{datasets}

