% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_en}
\alias{stopwords_en}
\title{English stopwords}
\format{An object of class \code{character} of length 174.}
\source{
\link[tm]{stopwords}
}
\usage{
stopwords_en
}
\description{
174 words frequenty used English words for removal in text mining,
and based on the \code{tm} package
}
\keyword{datasets}

