% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twtr_get_hashtags}
\alias{twtr_get_hashtags}
\title{twtr_get_hashtags}
\usage{
twtr_get_hashtags(x)
}
\arguments{
\item{\code{x}}{a vector of tweets}
}
\value{
a list of data frames about the hash tags
}
\description{
Extract hash tags and analyze them
}

