% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twtr_get_mentions}
\alias{twtr_get_mentions}
\title{Extract and analyze mentions from tweets}
\usage{
twtr_get_mentions(x)
}
\arguments{
\item{\code{x}}{a vector of tweets}
}
\value{
a list of data frames about the hash tags
\enumerate{
  \item \code{mentions}: a matrix of all mentions used, each row corresponding to a
  tweet.
  \item \code{top_mentions}: a data.frame of the top mentions, showing the frequency
  of usage, and sorted in descending order.
  \item \code{mention_count}: a data.frame with one column, where each row shows the
  number of mentions used in the corresponding tweet.
  \item \code{mention_freq}: a data.frame showing the number of tweets where one mention
  was used, where two mentions were used, etc
  \item \code{mention_summary}: a summary data.frame shwoing the number of tweets,
  total mention count, unique mentions, and mentions per tweet
}
}
\description{
Get the mentions from a vector of tweets, together with summaries about the
usage of those mentions; frequency, occurrences, percentages, and more
}
\examples{

tweets <- c("this is my first #tweet with #two hashtags and one @mention",
"my second tweet has no hashtags and no mentions",
"my third #tweet #has #the most hashtags and @mentions @mention")
twtr_get_mentions(tweets)
}

