% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twtr_get_urls}
\alias{twtr_get_urls}
\title{Extract and analyze URLs from tweets}
\usage{
twtr_get_urls(x)
}
\arguments{
\item{x}{a vector of tweets}
}
\value{
a list of data frames about the URLs
\enumerate{
  \item \code{urls}: a matrix of occurences of the URLs, each row corresponding to a
  tweet.
  \item \code{top_urls} a data frame of the most frequently used URLs, sorted
  \item \code{url_count}: a data frame with one column, where each row shows the
  number of times the URL was used in the corresponding tweet.
  \item \code{url_freq}: a data frame showing the number of tweets where the URL
  was used once, where the URL was used twice, etc
  \item \code{url_summary}: a summary data frame shwoing the number of tweets,
  total URL count, unique URLs, and URLs per tweet
}
}
\description{
Get URLs from a vector of tweets, together with summaries about their
usage; frequency, occurrences, percentages, and more
}
\examples{

tweets <- c("this is my first #tweet with #two hashtags and one @mention http://t.co/foo",
"my second tweet has no hashtags and no mentions https://t.co/bar",
"my third #tweet #has #the most hashtags and @mentions @mention http://sub.domain/something")
twtr_get_urls(tweets)
}

