% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twtr_get_word}
\alias{twtr_get_word}
\title{Extract and analyze words from tweets}
\usage{
twtr_get_word(x, word = "word", exact = TRUE)
}
\arguments{
\item{\code{x}}{a vector of tweets}

\item{\code{word}}{the word you want to extract}

\item{\code{exact}}{logical, do you want the \code{word} to be exactly matched
or do you want anny occurence of this sequence of characters?}
}
\value{
a list of data frames about the word
\enumerate{
  \item \code{words}: a matrix of occurences of the word, each row corresponding to a
  tweet.
  \item \code{word_count}: a data.frame with one column, where each row shows the
  number of times the word was used in the corresponding tweet.
  \item \code{word_freq}: a data.frame showing the number of tweets where the word
  was used once, where the word was used twice, etc
  \item \code{word_summary}: a summary data.frame shwoing the number of tweets,
  total word count, unique words, and words per tweet
}
}
\description{
Get a word from a vector of tweets, together with summaries about the
usage of this word; frequency, occurrences, percentages, and more
}
\examples{

tweets <- c("this is my first #tweet with #two hashtags and one @mention",
"my second tweet has no hashtags and no mentions",
"my third #tweet #has #the most hashtags and @mentions @mention")
twtr_get_word(tweets, word = "hashtags", exact = TRUE)
}

