% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twtr_get_words}
\alias{twtr_get_words}
\title{Extract and analyze words from tweets}
\usage{
twtr_get_words(x, words, exact = TRUE)
}
\arguments{
\item{x}{a vector of tweets}

\item{words}{a character vector of one or more words}

\item{exact}{logical, defaults to TRUE, do you want the \code{words}
to be exactly matched as complete words or do you want anny occurence of
this sequence of characters?}
}
\value{
a list of data frames about the word
\enumerate{
  \item \code{words}: a matrix of occurences of the word, each row corresponding to a
  tweet.
  \item \code{top_words} a data frame of the most frequently used words, sorted
  \item \code{word_count}: a data.frame with one column, where each row shows the
  number of times the word was used in the corresponding tweet.
  \item \code{word_freq}: a data.frame showing the number of tweets where the word
  was used once, where the word was used twice, etc
  \item \code{word_summary}: a summary data.frame shwoing the number of tweets,
  total word count, unique words, and words per tweet
}
}
\description{
Get a word from a vector of tweets, together with summaries about the
usage of this word; frequency, occurrences, percentages, and more
}
\examples{

tweets <- c("this is my first #tweet with #two hashtags and one @mention",
"my second tweet has no hashtags and no mentions",
"my third #tweet #has #the most hashtags and @mentions @mention")
twtr_get_words(tweets, word = "hashtags", exact = TRUE)
}

