% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_frequency.R
\name{word_frequency}
\alias{word_frequency}
\title{Word frequency}
\usage{
word_frequency(df)
}
\arguments{
\item{df}{a data frame with a character and a numeric vector columns}
}
\value{
a data frame of metrics by word along side the original metrics
}
\description{
Analyze the weighted frequency of words in a certain text with metrics.
Get hidden patterns in the text by extracting and aggregating metrics about words
in your text and comparing it to the original metric.
}
\details{
useful for:
- keyword / search term analysis
- tweets analysis
- articles traffic (from titles or URLs)
- Facebook statuses
- any other social media text with metrics
}
\examples{
word_frequency(boxoffice)
}

