% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/word_frequency.R
\docType{data}
\name{boxoffice}
\alias{boxoffice}
\title{Boxoffice revenues of 14,196 US movies}
\format{A data.frame with 14,196 observations and 5 variables
\describe{
  \item{rank}{ranking, based on alltime revenue}
  \item{title}{the name of the movie}
  \item{studio}{the production studio}
  \item{lifetime_gross}{lifetime boxoffice revenue}
  \item{year}{year of production}
}}
\source{
\href{http://www.boxofficemojo.com/alltime/domestic.htm}{Boxoffice Mojo}

\url{http://www.boxofficemojo.com/alltime/domestic.htm}
}
\usage{
boxoffice

boxoffice
}
\description{
A dataset of 14k movies produced since 1920, together with
alltime boxoffice revenues, as of June 2016.

A dataset of the top 100 grossing movies of all time, as of April
2016.
}
\keyword{datasets}

