% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{kw_broad}
\alias{kw_broad}
\title{Broad match keywords}
\usage{
kw_broad(x)
}
\arguments{
\item{x}{a keyword or a vector of keywords}
}
\value{
keywords without ", [, ], +, or -
}
\description{
This is useful in case you already have keywords as phrase, or exact for
example and want to return to broad. Another useful thing about this function is
that it serves as an intermediary state between transforming a keyword from phrase to exact
for example, it makes sure that you don't end up with something like ["keyword"].
}
\examples{
iris$exact <- kw_exact(iris$Species)
iris$broad <- kw_broad(iris$exact)
mtcars$phrase <- kw_phrase(rownames(mtcars))
mtcars$broad <- kw_broad(mtcars$phrase)
}
\keyword{AdWords,}
\keyword{Ads,}
\keyword{Bing}
\keyword{SEM,}
\keyword{marketing}
\keyword{online}

