% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{kw_modified_broad}
\alias{kw_modified_broad}
\title{Modified broad match keywords}
\usage{
kw_modified_broad(x)
}
\arguments{
\item{x}{a keyword or a vector of keywords}
}
\value{
modified broad match keywords
}
\description{
Add a plus sign before each word of the keyword / keyphrase.
Note that sometimes it's better to only have one or two words with the plus sign
next to them.
}
\examples{
keywords <- c("chocolate stores", "buy chocolate online", "best chocolate")
kw_modified_broad(keywords)
}

