% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp.R
\name{serp_google}
\alias{serp_google}
\title{serp_google}
\usage{
serp_google(query, cse_id, api_key, lang = "en", user_loc = "us",
  goog_domain = "com")
}
\arguments{
\item{query}{the keyword you want to research}

\item{cse_id}{your google custom search engine ID}

\item{api_key}{your API key}

\item{lang}{a two-letter representation of the user langugae}

\item{user_loc}{location of the user given with the two-letter ISO code}

\item{goog_domain}{the domain of Google which you want it to search in
Simply enter a keyword, with optional parameters, and get the page of search results returned as a data.frame.}
}
\description{
Search engine results page.
}

