% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twtr_get_mentions}
\alias{twtr_get_mentions}
\title{twtr_get_mentions}
\usage{
twtr_get_mentions(x)
}
\arguments{
\item{\code{x}}{a vector of tweets}
}
\value{
a list of data frames about the mentions
}
\description{
Extract mentions and analyze them
}

